+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/services/github/GitHubTask.java
@@ -39,8 +39,12 @@ class GitHubTask implements Runnable {
 
     @Override
     public void run() {
-        gitHubService.getConnectors().forEach(GitHubConnector::download);
+        connectAndCache();
         grabTextures();
+    }
+
+    private void connectAndCache() {
+        gitHubService.getConnectors().forEach(GitHubConnector::download);
     }
 
     /**