+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/abstractItems/AGenerator.java
@@ -152,7 +152,7 @@ public abstract class AGenerator extends AbstractEnergyProvider {
                 ChestMenuUtils.updateProgressbar(inv, 22, timeleft, processing.get(l).getTicks(), getProgressBar());
 
                 if (isChargeable()) {
-                    int charge = getCharge(l);
+                    int charge = getCharge(l, data);
 
                     if (getCapacity() - charge >= getEnergyProduction()) {
                         progress.put(l, timeleft - 1);